requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
openai==1.8.0
python-dotenv==1.0.1
